move dashboard components to dedicated subfolder
