+++ components/dashboard_template/app-sidebar.tsx
@@ -2,10 +2,10 @@
 
 import * as React from "react"
 
-import { NavDocuments } from "@/components/nav-documents"
-import { NavMain } from "@/components/nav-main"
-import { NavSecondary } from "@/components/nav-secondary"
-import { NavUser } from "@/components/nav-user"
+import { NavDocuments } from "@/components/dashboard_template/nav-documents"
+import { NavMain } from "@/components/dashboard_template/nav-main"
+import { NavSecondary } from "@/components/dashboard_template/nav-secondary"
+import { NavUser } from "@/components/dashboard_template/nav-user"
 import {
   Sidebar,
   SidebarContent,
