+++ components/dashboard_template/chart-area-interactive.tsx
